refactor(simulation): extract compiled helpers
