+++ beamz/simulation/compiled.py
@@ -32,6 +32,7 @@ from beamz.devices.sources.compiler import (
     compile_source_specs,
 )
 from beamz.simulation import ops
+from beamz.simulation import shell, sources, track
 from beamz.simulation.material_models import (
     CompiledMaterialSpec,
     MaterialState,
@@ -229,11 +230,7 @@ class CompiledSimulation:
     def _sources_for(
         self, timing: str, component: str
     ) -> tuple[CompiledSourceSpec, ...]:
-        return tuple(
-            s
-            for s in self.source_specs
-            if s.timing == timing and s.component == component
-        )
+        return sources.sources_for(self, timing, component)
 
     def _apply_specs(
         self,
@@ -241,21 +238,7 @@ class CompiledSimulation:
         abs_step: jnp.ndarray,
         specs: tuple[CompiledSourceSpec, ...],
     ) -> jnp.ndarray:
-        out = arr
-        for spec in specs:
-            safe_idx = jnp.clip(abs_step, 0, spec.waveform.shape[0] - 1)
-            amp = spec.waveform[safe_idx]
-            if (
-                spec.is_slab
-                and spec.slab_starts is not None
-                and spec.slab_sizes is not None
-            ):
-                patch = spec.coeff * amp
-                cur = jax.lax.dynamic_slice(out, spec.slab_starts, spec.slab_sizes)
-                out = jax.lax.dynamic_update_slice(out, cur + patch, spec.slab_starts)
-            else:
-                out = out.at[spec.index].add(spec.coeff * amp)
-        return out
+        return sources.apply_specs(self, arr, abs_step, specs)
 
     def _apply_batched_slabs(
         self,
@@ -263,51 +246,7 @@ class CompiledSimulation:
         abs_step: jnp.ndarray,
         group: BatchedSlabGroup,
     ) -> jnp.ndarray:
-        """Apply stacked slab sources via fori_loop (constant HLO size)."""
-        safe_idx = jnp.clip(abs_step, 0, group.waveforms.shape[1] - 1)
-        ndim = len(group.max_sizes)
-
-        # Common hot path for benchmarks and many production setups:
-        # one slab source only (e.g., single Gaussian source). Avoid nested
-        # fori_loop overhead in the timestep kernel.
-        if group.n == 1:
-            amp = group.waveforms[0, safe_idx]
-            starts_0 = group.starts_tuple[0]
-            if self.config.source_single_slab_dense:
-                # Optional DUS-free path for diagnostics: materialize a full-grid
-                # coefficient tensor once and inject via dense add.
-                pad_width = tuple(
-                    (
-                        starts_0[d],
-                        int(arr.shape[d]) - starts_0[d] - group.max_sizes[d],
-                    )
-                    for d in range(ndim)
-                )
-                dense_coeff = jnp.pad(group.coeffs[0], pad_width)
-                return arr + dense_coeff * amp
-            patch = group.coeffs[0] * amp
-            cur = jax.lax.dynamic_slice(arr, starts_0, group.max_sizes)
-            return jax.lax.dynamic_update_slice(arr, cur + patch, starts_0)
-
-        if group.n == 2:
-
-            def apply_one(out, i: int):
-                amp_i = group.waveforms[i, safe_idx]
-                patch_i = group.coeffs[i] * amp_i
-                starts_i = group.starts_tuple[i]
-                cur_i = jax.lax.dynamic_slice(out, starts_i, group.max_sizes)
-                return jax.lax.dynamic_update_slice(out, cur_i + patch_i, starts_i)
-
-            return apply_one(apply_one(arr, 0), 1)
-
-        def body(i, out):
-            amp = group.waveforms[i, safe_idx]
-            patch = group.coeffs[i] * amp
-            starts_i = [group.starts[i, d] for d in range(ndim)]
-            cur = jax.lax.dynamic_slice(out, starts_i, group.max_sizes)
-            return jax.lax.dynamic_update_slice(out, cur + patch, starts_i)
-
-        return jax.lax.fori_loop(0, group.n, body, arr)
+        return sources.apply_batched_slabs(self, arr, abs_step, group)
 
     def _apply_source_group(
         self,
@@ -316,12 +255,7 @@ class CompiledSimulation:
         batch: BatchedSlabGroup | None,
         rest: tuple[CompiledSourceSpec, ...],
     ) -> jnp.ndarray:
-        """Apply batched slab sources then remaining non-slab sources."""
-        if batch is not None:
-            arr = self._apply_batched_slabs(arr, abs_step, batch)
-        if rest:
-            arr = self._apply_specs(arr, abs_step, rest)
-        return arr
+        return sources.apply_group(self, arr, abs_step, batch, rest)
 
     def _apply_lossy_shell(
         self,
@@ -332,16 +266,7 @@ class CompiledSimulation:
         source: jnp.ndarray,
         slabs: tuple[tuple[tuple[int, int, int], tuple[int, int, int]], ...],
     ) -> jnp.ndarray:
-        """Apply lossy E update only on precomputed disjoint boundary slabs."""
-        out = updated
-        for starts, sizes in slabs:
-            old_s = jax.lax.dynamic_slice(old, starts, sizes)
-            curl_s = jax.lax.dynamic_slice(curl, starts, sizes)
-            decay_s = jax.lax.dynamic_slice(decay, starts, sizes)
-            source_s = jax.lax.dynamic_slice(source, starts, sizes)
-            lossy_s = decay_s * old_s + source_s * curl_s
-            out = jax.lax.dynamic_update_slice(out, lossy_s, starts)
-        return out
+        return shell.apply_lossy_shell(self, updated, old, curl, decay, source, slabs)
 
     def _apply_lossy_shell_from_lossless(
         self,
@@ -352,26 +277,15 @@ class CompiledSimulation:
         source_lossless: jnp.ndarray,
         slabs: tuple[tuple[tuple[int, int, int], tuple[int, int, int]], ...],
     ) -> jnp.ndarray:
-        """Apply lossy correction on shell slabs using old/lossless fields only.
-
-        Given:
-            lossless = old +/- source_lossless * curl
-            lossy    = decay * old +/- source * curl
-        we can eliminate curl on the shell via:
-            beta = source / source_lossless
-            lossy = (decay - beta) * old + beta * lossless
-        """
-        out = updated_lossless
-        for starts, sizes in slabs:
-            old_s = jax.lax.dynamic_slice(old, starts, sizes)
-            lossless_s = jax.lax.dynamic_slice(updated_lossless, starts, sizes)
-            decay_s = jax.lax.dynamic_slice(decay, starts, sizes)
-            source_s = jax.lax.dynamic_slice(source, starts, sizes)
-            source_ll_s = jax.lax.dynamic_slice(source_lossless, starts, sizes)
-            beta = source_s / source_ll_s
-            lossy_s = (decay_s - beta) * old_s + beta * lossless_s
-            out = jax.lax.dynamic_update_slice(out, lossy_s, starts)
-        return out
+        return shell.apply_lossy_shell_from_lossless(
+            self,
+            updated_lossless,
+            old,
+            decay,
+            source,
+            source_lossless,
+            slabs,
+        )
 
     def _monitor_power_2d(
         self,
@@ -380,15 +294,7 @@ class CompiledSimulation:
         hx: jnp.ndarray,
         hy: jnp.ndarray,
     ) -> jnp.ndarray:
-        power_scale = jnp.asarray(spec.power_scale, dtype=jnp.float32)
-        ez_vals = ez[spec.y_ez, spec.x_ez] * spec.valid_ez
-        hx_vals = hx[spec.y_hx, spec.x_hx] * spec.valid_hx
-        hy_vals = hy[spec.y_hy, spec.x_hy] * spec.valid_hy
-
-        sx = -ez_vals * hy_vals
-        sy = ez_vals * hx_vals
-        mag = jnp.sqrt(sx * sx + sy * sy)
-        return jnp.asarray(jnp.sum(mag), dtype=jnp.float32) * power_scale
+        return track.monitor_power_2d(self, spec, ez, hx, hy)
 
     def _monitor_power_3d(
         self,
@@ -400,19 +306,7 @@ class CompiledSimulation:
         hy: jnp.ndarray,
         hz: jnp.ndarray,
     ) -> jnp.ndarray:
-        power_scale = jnp.asarray(spec.power_scale, dtype=jnp.float32)
-        exs = ex[spec.ex_idx][: spec.min_dim0, : spec.min_dim1]
-        eys = ey[spec.ey_idx][: spec.min_dim0, : spec.min_dim1]
-        ezs = ez[spec.ez_idx][: spec.min_dim0, : spec.min_dim1]
-        hxs = hx[spec.hx_idx][: spec.min_dim0, : spec.min_dim1]
-        hys = hy[spec.hy_idx][: spec.min_dim0, : spec.min_dim1]
-        hzs = hz[spec.hz_idx][: spec.min_dim0, : spec.min_dim1]
-
-        sx = eys * hzs - ezs * hys
-        sy = ezs * hxs - exs * hzs
-        sz = exs * hys - eys * hxs
-        mag = jnp.sqrt(sx * sx + sy * sy + sz * sz)
-        return jnp.asarray(jnp.sum(mag), dtype=jnp.float32) * power_scale
+        return track.monitor_power_3d(self, spec, ex, ey, ez, hx, hy, hz)
 
     def _update_monitors(
         self,
@@ -429,289 +323,20 @@ class CompiledSimulation:
         batched_mon: BatchedMonitorData | None = None,
         monitors_2d: tuple[CompiledMonitorSpec, ...] = (),
     ) -> MonitorState:
-        if not self.monitor_specs:
-            return monitor_state
-
-        powers = monitor_state.powers
-        timestamps = monitor_state.timestamps
-        counts = monitor_state.counts
-        freq_flux_re = monitor_state.freq_flux_re
-        freq_flux_im = monitor_state.freq_flux_im
-        freq_phase_re = monitor_state.freq_phase_re
-        freq_phase_im = monitor_state.freq_phase_im
-        dft_vec_re = monitor_state.dft_vec_re
-        dft_vec_im = monitor_state.dft_vec_im
-        dft_weight_sum = monitor_state.dft_weight_sum
-        max_records = powers.shape[1]
-
-        # Batched 3D monitors via fori_loop (constant HLO size)
-        if batched_mon is not None:
-            bm = batched_mon
-            ex_flat = ex.ravel()
-            ey_flat = ey.ravel()
-            ez_flat = ez.ravel()
-            hx_flat = hx.ravel()
-            hy_flat = hy.ravel()
-            hz_flat = hz.ravel()
-
-            def _mon_body(i, carry):
-                pwr, ts, cnt, f_re, f_im, ph_re, ph_im, d_re, d_im, d_w = carry
-                mi = bm.monitor_indices[i]
-
-                should_record = (abs_step % bm.record_intervals[i]) == 0
-                can_record = cnt[mi] < max_records
-                do_record = should_record & can_record & bm.accumulate_flags[i]
-
-                mask = bm.valid_mask[i]
-                exs = ex_flat[bm.ex_flat_idx[i]] * mask
-                eys = ey_flat[bm.ey_flat_idx[i]] * mask
-                ezs = ez_flat[bm.ez_flat_idx[i]] * mask
-                hxs = hx_flat[bm.hx_flat_idx[i]] * mask
-                hys = hy_flat[bm.hy_flat_idx[i]] * mask
-                hzs = hz_flat[bm.hz_flat_idx[i]] * mask
-
-                sx = eys * hzs - ezs * hys
-                sy = ezs * hxs - exs * hzs
-                sz = exs * hys - eys * hxs
-                power_val = (
-                    jnp.sum(jnp.sqrt(sx * sx + sy * sy + sz * sz)) * bm.power_scales[i]
-                )
-                axis_i = bm.normal_axes[i]
-                normal_flux = (
-                    jnp.sum(jnp.where(axis_i == 0, sx, jnp.where(axis_i == 1, sy, sz)))
-                    * bm.power_scales[i]
-                )
-                flux_sample = jnp.where(axis_i < 0, power_val, normal_flux)
-
-                slot = jnp.minimum(cnt[mi], max_records - 1)
-                pwr = pwr.at[mi, slot].set(
-                    jnp.where(do_record, power_val, pwr[mi, slot])
-                )
-                ts = ts.at[mi, slot].set(jnp.where(do_record, t_phys, ts[mi, slot]))
-                cnt = cnt.at[mi].set(cnt[mi] + jnp.where(do_record, 1, 0))
-                do_freq = bm.freq_enabled[i] & (
-                    (abs_step % bm.freq_record_intervals[i]) == 0
-                )
-                mask_f = bm.freq_mask[i]
-                row_f_re = f_re[mi]
-                row_f_im = f_im[mi]
-                row_ph_re = ph_re[mi]
-                row_ph_im = ph_im[mi]
-                cur_ph_re = row_ph_re
-                cur_ph_im = row_ph_im
-                delta_re = flux_sample * dt_scalar * row_ph_re * mask_f
-                delta_im = flux_sample * dt_scalar * row_ph_im * mask_f
-                row_f_re = row_f_re + jnp.where(do_freq, delta_re, 0.0)
-                row_f_im = row_f_im + jnp.where(do_freq, delta_im, 0.0)
-                rot_re = bm.freq_rot_re[i]
-                rot_im = bm.freq_rot_im[i]
-                next_ph_re = row_ph_re * rot_re - row_ph_im * rot_im
-                next_ph_im = row_ph_re * rot_im + row_ph_im * rot_re
-                row_ph_re = jnp.where(do_freq, next_ph_re, row_ph_re)
-                row_ph_im = jnp.where(do_freq, next_ph_im, row_ph_im)
-                f_re = f_re.at[mi].set(row_f_re)
-                f_im = f_im.at[mi].set(row_f_im)
-                ph_re = ph_re.at[mi].set(row_ph_re)
-                ph_im = ph_im.at[mi].set(row_ph_im)
-                do_dft = (
-                    bm.dft_enabled[i]
-                    & ((abs_step % bm.dft_record_intervals[i]) == 0)
-                    & (t_phys >= bm.dft_t_start[i])
-                    & (t_phys <= bm.dft_t_end[i])
-                )
-                two_pi = jnp.asarray(2.0 * np.pi, dtype=jnp.float32)
-                span = jnp.maximum(bm.dft_t_end[i] - bm.dft_t_start[i], 1e-30)
-                tau = jnp.clip((t_phys - bm.dft_t_start[i]) / span, 0.0, 1.0)
-                w_hann = 0.5 * (1.0 - jnp.cos(two_pi * tau))
-                w = jnp.where(
-                    bm.dft_window_code[i] == 1,
-                    w_hann,
-                    jnp.asarray(1.0, dtype=jnp.float32),
-                )
-                w = jnp.where(do_dft, w, jnp.asarray(0.0, dtype=jnp.float32))
-                ph_vec_re = cur_ph_re * mask_f
-                ph_vec_im = cur_ph_im * mask_f
-
-                vecs = jnp.stack((exs, eys, ezs, hxs, hys, hzs), axis=0)
-                comp_mask = bm.dft_component_mask[i][:, None, None]
-                delta_re_3d = w * comp_mask * jnp.einsum("f,cp->cfp", ph_vec_re, vecs)
-                delta_im_3d = w * comp_mask * jnp.einsum("f,cp->cfp", ph_vec_im, vecs)
-                d_re = d_re.at[mi].add(delta_re_3d)
-                d_im = d_im.at[mi].add(delta_im_3d)
-                d_w = d_w.at[mi].add(w * mask_f)
-
-                return pwr, ts, cnt, f_re, f_im, ph_re, ph_im, d_re, d_im, d_w
-
-            (
-                powers,
-                timestamps,
-                counts,
-                freq_flux_re,
-                freq_flux_im,
-                freq_phase_re,
-                freq_phase_im,
-                dft_vec_re,
-                dft_vec_im,
-                dft_weight_sum,
-            ) = jax.lax.fori_loop(
-                0,
-                bm.n_monitors,
-                _mon_body,
-                (
-                    powers,
-                    timestamps,
-                    counts,
-                    freq_flux_re,
-                    freq_flux_im,
-                    freq_phase_re,
-                    freq_phase_im,
-                    dft_vec_re,
-                    dft_vec_im,
-                    dft_weight_sum,
-                ),
-            )
-
-        # Remaining static monitor specs (2D or 3D).
-        for mon in monitors_2d:
-            should_record = (abs_step % mon.record_interval) == 0
-            can_record = counts[mon.monitor_index] < max_records
-            do_record = should_record & can_record & mon.accumulate_power
-            do_freq = (
-                (abs_step % mon.freq_record_interval) == 0
-                if mon.accumulate_frequency and mon.freq_count > 0
-                else jnp.array(False)
-            )
-            need_sample = do_record | do_freq
-
-            power_sample = jnp.where(
-                need_sample,
-                (
-                    self._monitor_power_3d(mon, ex, ey, ez, hx, hy, hz)
-                    if mon.is_3d
-                    else self._monitor_power_2d(mon, ez, hx, hy)
-                ),
-                jnp.array(0.0, dtype=jnp.float32),
-            )
-            power_val = jnp.where(
-                do_record, power_sample, jnp.array(0.0, dtype=jnp.float32)
-            )
-
-            slot = jnp.minimum(counts[mon.monitor_index], max_records - 1)
-            old_power = powers[mon.monitor_index, slot]
-            old_ts = timestamps[mon.monitor_index, slot]
-
-            powers = powers.at[mon.monitor_index, slot].set(
-                jnp.where(do_record, power_val, old_power)
-            )
-            timestamps = timestamps.at[mon.monitor_index, slot].set(
-                jnp.where(do_record, t_phys, old_ts)
-            )
-            counts = counts.at[mon.monitor_index].set(
-                counts[mon.monitor_index] + jnp.where(do_record, 1, 0)
-            )
-            if mon.accumulate_frequency and mon.freq_count > 0:
-                mi = mon.monitor_index
-                row_f_re = freq_flux_re[mi, : mon.freq_count]
-                row_f_im = freq_flux_im[mi, : mon.freq_count]
-                row_ph_re = freq_phase_re[mi, : mon.freq_count]
-                row_ph_im = freq_phase_im[mi, : mon.freq_count]
-                cur_ph_re = row_ph_re
-                cur_ph_im = row_ph_im
-                delta_re = power_sample * dt_scalar * row_ph_re
-                delta_im = power_sample * dt_scalar * row_ph_im
-                row_f_re = row_f_re + jnp.where(do_freq, delta_re, 0.0)
-                row_f_im = row_f_im + jnp.where(do_freq, delta_im, 0.0)
-                next_ph_re = row_ph_re * mon.freq_rot_re - row_ph_im * mon.freq_rot_im
-                next_ph_im = row_ph_re * mon.freq_rot_im + row_ph_im * mon.freq_rot_re
-                row_ph_re = jnp.where(do_freq, next_ph_re, row_ph_re)
-                row_ph_im = jnp.where(do_freq, next_ph_im, row_ph_im)
-                freq_flux_re = freq_flux_re.at[mi, : mon.freq_count].set(row_f_re)
-                freq_flux_im = freq_flux_im.at[mi, : mon.freq_count].set(row_f_im)
-                freq_phase_re = freq_phase_re.at[mi, : mon.freq_count].set(row_ph_re)
-                freq_phase_im = freq_phase_im.at[mi, : mon.freq_count].set(row_ph_im)
-            if mon.dft_enabled and mon.freq_count > 0 and mon.dft_point_count > 0:
-                mi = mon.monitor_index
-                if mon.accumulate_frequency and mon.freq_count > 0:
-                    dft_ph_re = cur_ph_re
-                    dft_ph_im = cur_ph_im
-                else:
-                    dft_ph_re = freq_phase_re[mi, : mon.freq_count]
-                    dft_ph_im = freq_phase_im[mi, : mon.freq_count]
-                do_dft = (
-                    ((abs_step % mon.dft_record_interval) == 0)
-                    & (t_phys >= mon.dft_t_start)
-                    & (t_phys <= mon.dft_t_end)
-                )
-                two_pi = jnp.asarray(2.0 * np.pi, dtype=jnp.float32)
-                span = jnp.maximum(
-                    mon.dft_t_end - mon.dft_t_start,
-                    jnp.asarray(1e-30, dtype=jnp.float32),
-                )
-                tau = jnp.asarray(
-                    jnp.clip((t_phys - mon.dft_t_start) / span, 0.0, 1.0),
-                    dtype=jnp.float32,
-                )
-                w_hann = 0.5 * (1.0 - jnp.cos(two_pi * tau))
-                w = jnp.asarray(
-                    jnp.where(
-                        mon.dft_window_code == 1,
-                        w_hann,
-                        jnp.asarray(1.0, dtype=jnp.float32),
-                    ),
-                    dtype=jnp.float32,
-                )
-                w = jnp.asarray(
-                    jnp.where(do_dft, w, jnp.asarray(0.0, dtype=jnp.float32)),
-                    dtype=jnp.float32,
-                )
-
-                if mon.is_3d:
-                    ex_vals = ex[mon.ex_idx][: mon.min_dim0, : mon.min_dim1].reshape(-1)
-                    ey_vals = ey[mon.ey_idx][: mon.min_dim0, : mon.min_dim1].reshape(-1)
-                    ez_vals = ez[mon.ez_idx][: mon.min_dim0, : mon.min_dim1].reshape(-1)
-                    hx_vals = hx[mon.hx_idx][: mon.min_dim0, : mon.min_dim1].reshape(-1)
-                    hy_vals = hy[mon.hy_idx][: mon.min_dim0, : mon.min_dim1].reshape(-1)
-                    hz_vals = hz[mon.hz_idx][: mon.min_dim0, : mon.min_dim1].reshape(-1)
-                else:
-                    ex_vals = ex[mon.y_ex, mon.x_ex] * mon.valid_ex
-                    ey_vals = ey[mon.y_ey, mon.x_ey] * mon.valid_ey
-                    ez_vals = ez[mon.y_ez, mon.x_ez] * mon.valid_ez
-                    hx_vals = hx[mon.y_hx, mon.x_hx] * mon.valid_hx
-                    hy_vals = hy[mon.y_hy, mon.x_hy] * mon.valid_hy
-                    hz_vals = hz[mon.y_hz, mon.x_hz] * mon.valid_hz
-                vecs = jnp.stack(
-                    (ex_vals, ey_vals, ez_vals, hx_vals, hy_vals, hz_vals), axis=0
-                )
-                comp_mask = mon.dft_component_mask[:, None, None]
-                delta_re = jnp.asarray(
-                    w * comp_mask * jnp.einsum("f,cp->cfp", dft_ph_re, vecs),
-                    dtype=jnp.float32,
-                )
-                delta_im = jnp.asarray(
-                    w * comp_mask * jnp.einsum("f,cp->cfp", dft_ph_im, vecs),
-                    dtype=jnp.float32,
-                )
-                dft_vec_re = dft_vec_re.at[
-                    mi, :, : mon.freq_count, : mon.dft_point_count
-                ].add(delta_re[:, : mon.freq_count, : mon.dft_point_count])
-                dft_vec_im = dft_vec_im.at[
-                    mi, :, : mon.freq_count, : mon.dft_point_count
-                ].add(delta_im[:, : mon.freq_count, : mon.dft_point_count])
-                dft_weight_sum = dft_weight_sum.at[mi, : mon.freq_count].add(
-                    jnp.asarray(w, dtype=jnp.float32)
-                )
-
-        return MonitorState(
-            powers=powers,
-            timestamps=timestamps,
-            counts=counts,
-            freq_flux_re=freq_flux_re,
-            freq_flux_im=freq_flux_im,
-            freq_phase_re=freq_phase_re,
-            freq_phase_im=freq_phase_im,
-            dft_vec_re=dft_vec_re,
-            dft_vec_im=dft_vec_im,
-            dft_weight_sum=dft_weight_sum,
+        return track.update_monitors(
+            self,
+            monitor_state,
+            abs_step,
+            t_phys,
+            dt_scalar,
+            ex,
+            ey,
+            ez,
+            hx,
+            hy,
+            hz,
+            batched_mon=batched_mon,
+            monitors_2d=monitors_2d,
         )
 
     def _build_scan(self):
@@ -1183,130 +808,30 @@ class CompiledSimulation:
 
     def apply_monitor_state(self, monitor_state: MonitorState):
         """Push monitor-state buffers back to Monitor objects."""
-        for spec in self.monitor_specs:
-            dev = self.monitor_devices[spec.monitor_index]
-            count = int(np.asarray(monitor_state.counts[spec.monitor_index]))
-            powers = np.asarray(
-                monitor_state.powers[spec.monitor_index, :count], dtype=float
-            )
-            ts = np.asarray(
-                monitor_state.timestamps[spec.monitor_index, :count], dtype=float
-            )
-
-            dev.power_history = list(powers.tolist())
-            dev.power_timestamps = list(ts.tolist())
-            dev.power_accumulation_count = count
-            if spec.freq_count > 0:
-                re = np.asarray(
-                    monitor_state.freq_flux_re[spec.monitor_index, : spec.freq_count],
-                    dtype=np.float32,
-                )
-                im = np.asarray(
-                    monitor_state.freq_flux_im[spec.monitor_index, : spec.freq_count],
-                    dtype=np.float32,
-                )
-                dev.frequency_flux_spectrum = (re + 1j * im).astype(np.complex64)
-            else:
-                dev.frequency_flux_spectrum = np.zeros((0,), dtype=np.complex64)
-
-            if spec.dft_enabled and spec.freq_count > 0 and spec.dft_point_count > 0:
-                comp_names = ("Ex", "Ey", "Ez", "Hx", "Hy", "Hz")
-                comp_mask = (
-                    np.asarray(spec.dft_component_mask, dtype=np.float32)
-                    if spec.dft_component_mask is not None
-                    else np.ones((6,), dtype=np.float32)
-                )
-                weight_sum = np.asarray(
-                    monitor_state.dft_weight_sum[spec.monitor_index, : spec.freq_count],
-                    dtype=np.float64,
-                )
-                dev._dft_weight_sum = weight_sum
-                dev._dft_accum = {}
-                for comp_i, comp_name in enumerate(comp_names):
-                    if comp_mask[comp_i] <= 0.0:
-                        continue
-                    re = np.asarray(
-                        monitor_state.dft_vec_re[
-                            spec.monitor_index,
-                            comp_i,
-                            : spec.freq_count,
-                            : spec.dft_point_count,
-                        ],
-                        dtype=np.float64,
-                    )
-                    im = np.asarray(
-                        monitor_state.dft_vec_im[
-                            spec.monitor_index,
-                            comp_i,
-                            : spec.freq_count,
-                            : spec.dft_point_count,
-                        ],
-                        dtype=np.float64,
-                    )
-                    dev._dft_accum[comp_name] = re + 1j * im
-            else:
-                dev._dft_weight_sum = np.zeros((0,), dtype=np.float64)
-                dev._dft_accum = {}
+        track.apply_monitor_state(self, monitor_state)
 
 
 def monitor_state_size(specs: tuple[CompiledMonitorSpec, ...], num_steps: int) -> int:
-    if not specs:
-        return 0
-    return int(
-        max(
-            int(np.ceil(num_steps / max(1, int(spec.record_interval))))
-            for spec in specs
-        )
-    )
+    return track.monitor_state_size(specs, num_steps)
 
 
 def monitor_frequency_size(specs: tuple[CompiledMonitorSpec, ...]) -> int:
-    if not specs:
-        return 0
-    return int(max(int(spec.freq_count) for spec in specs))
+    return track.monitor_frequency_size(specs)
 
 
 def monitor_dft_point_size(specs: tuple[CompiledMonitorSpec, ...]) -> int:
-    if not specs:
-        return 0
-    return int(max(int(getattr(spec, "dft_point_count", 0)) for spec in specs))
+    return track.monitor_dft_point_size(specs)
 
 
 def _edge_full_thickness(mask: np.ndarray, axis: int) -> tuple[int, int]:
-    """Count leading/trailing planes that are fully lossy along a given axis."""
-    other_axes = tuple(i for i in range(mask.ndim) if i != axis)
-    plane_all = mask.all(axis=other_axes)
-
-    left = 0
-    n = plane_all.shape[0]
-    while left < n and bool(plane_all[left]):
-        left += 1
-
-    right = 0
-    while right < (n - left) and bool(plane_all[n - 1 - right]):
-        right += 1
-    return left, right
+    return shell.edge_full_thickness(mask, axis)
 
 
 def _region_offsets_and_sizes(
     field_shape: tuple[int, ...],
     region: tuple[slice, ...],
 ) -> tuple[tuple[int, ...], tuple[int, ...]] | None:
-    """Return region starts/sizes for slice-only regions with unit strides."""
-    if len(field_shape) != len(region):
-        return None
-
-    starts: list[int] = []
-    sizes: list[int] = []
-    for dim, key in zip(field_shape, region):
-        if not isinstance(key, slice):
-            return None
-        start, stop, step = key.indices(dim)
-        if step != 1:
-            return None
-        starts.append(int(start))
-        sizes.append(int(max(stop - start, 0)))
-    return tuple(starts), tuple(sizes)
+    return shell.region_offsets_and_sizes(field_shape, region)
 
 
 def _infer_lossy_shell_slabs(
@@ -1314,73 +839,7 @@ def _infer_lossy_shell_slabs(
     region: tuple[slice, ...],
     conductivity_region: jnp.ndarray,
 ) -> tuple[bool, tuple[tuple[tuple[int, int, int], tuple[int, int, int]], ...]]:
-    """Infer disjoint boundary-shell slabs from conductivity mask.
-
-    Returns (enabled, slabs). Enabled is True only if the lossy mask can be represented as
-    a standard disjoint shell decomposition in 3D.
-    """
-    if len(field_shape) != 3:
-        return False, tuple()
-
-    region_layout = _region_offsets_and_sizes(field_shape, region)
-    if region_layout is None:
-        return False, tuple()
-    region_starts, region_sizes = region_layout
-    if len(region_sizes) != 3 or any(s <= 0 for s in region_sizes):
-        return False, tuple()
-
-    local_mask = np.asarray(conductivity_region) > 0.0
-    if tuple(local_mask.shape) != tuple(region_sizes):
-        return False, tuple()
-    if not local_mask.any():
-        return False, tuple()
-
-    zL, zR = _edge_full_thickness(local_mask, axis=0)
-    yL, yR = _edge_full_thickness(local_mask, axis=1)
-    xL, xR = _edge_full_thickness(local_mask, axis=2)
-
-    nz, ny, nx = region_sizes
-    z0, z1 = zL, nz - zR
-    y0, y1 = yL, ny - yR
-    x0, x1 = xL, nx - xR
-    if z0 > z1 or y0 > y1 or x0 > x1:
-        return False, tuple()
-
-    slabs: list[tuple[tuple[int, int, int], tuple[int, int, int]]] = []
-
-    def add_slab(starts: tuple[int, int, int], sizes: tuple[int, int, int]):
-        if all(s > 0 for s in sizes):
-            slabs.append((starts, sizes))
-
-    # Disjoint shell: z faces, then y faces within z-core, then x faces within yz-core.
-    add_slab((0, 0, 0), (zL, ny, nx))
-    add_slab((z1, 0, 0), (zR, ny, nx))
-    add_slab((z0, 0, 0), (max(z1 - z0, 0), yL, nx))
-    add_slab((z0, y1, 0), (max(z1 - z0, 0), yR, nx))
-    add_slab((z0, y0, 0), (max(z1 - z0, 0), max(y1 - y0, 0), xL))
-    add_slab((z0, y0, x1), (max(z1 - z0, 0), max(y1 - y0, 0), xR))
-
-    if not slabs:
-        return False, tuple()
-
-    recon = np.zeros(region_sizes, dtype=bool)
-    for starts, sizes in slabs:
-        z, y, x = starts
-        dz, dy, dx = sizes
-        recon[z : z + dz, y : y + dy, x : x + dx] = True
-
-    if not np.array_equal(recon, local_mask):
-        return False, tuple()
-
-    z_off, y_off, x_off = region_starts
-    global_slabs = tuple(
-        (
-            (starts[0] + z_off, starts[1] + y_off, starts[2] + x_off),
-            sizes,
-        )
-        for starts, sizes in slabs
-    )
-    return True, global_slabs
+    return shell.infer_lossy_shell_slabs(field_shape, region, conductivity_region)
 
 
 def _lossy_fraction(
@@ -1388,10 +847,7 @@ def _lossy_fraction(
     region: tuple[slice, ...],
     conductivity_region: jnp.ndarray,
 ) -> float:
-    """Fraction of lossy voxels for a field component."""
-    full_mask = np.zeros(field_shape, dtype=bool)
-    full_mask[region] = np.asarray(conductivity_region) > 0.0
-    return float(full_mask.mean())
+    return shell.lossy_fraction(field_shape, region, conductivity_region)
 
 
 def compile_simulation(design, devices, boundaries, run_cfg) -> CompiledSimulation:
